requirements_without_versions.txt
appnope
argon2-cffi
argon2-cffi-bindings
asttokens
attrs
backcall
beautifulsoup4
bleach
Brotli
cachetools
certifi
cffi
charset-normalizer
click
dash
dash-bootstrap-components
dash-core-components
dash-daq
dash-html-components
dash-table
db-dtypes
debugpy
decorator
defusedxml
entrypoints
executing
fastjsonschema
Flask
Flask-Compress
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kaleido
lxml
MarkupSafe
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
oauthlib
packaging
pandas
pandas-gbq
pandocfilters
parso
pexpect
pickleshare
plotly
prometheus-client
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycparser
pydata-google-auth
Pygments
pyparsing
pyrsistent
pyserial
python-dateutil
pytz
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
rsa
Send2Trash
six
soupsieve
stack-data
tenacity
terminado
tinycss2
tornado
traitlets
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
zipp
